refactor(tui): move config.set handler into methods_config (server.py god-file slice R4)
